May 12 Design Ideas
Please read through it. There are a couple sections I added, which I put
in bold. As well as some questions on how we want to do some things.
email me/Skype message me with ideas you have or feel free to update the
doc yourself. Nothing on here is finalized these are just ideas so feel
free to put your input in.